Trabalho

+++ 2DI_1190974_1191098/asm.h
@@ -2,6 +2,6 @@
 #define ASM_H
 
 short int extremas(int temp);
-int stats(int *temps, unsigned int dias, int *max_min);
+int stats(int *temps, unsigned int dias, short int *max_min);
 
 #endif
